fix .net framework build
