+++ Snowflake.Data.Tests/IntegrationTests/ChunkDownloadCancellationTest.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Snowflake.Data.Client;
 using Snowflake.Data.Core;
 using Snowflake.Data.Tests.Util;
+using Snowflake.Data.Tests.Util.Shims;
 using Xunit;
 
 namespace Snowflake.Data.Tests.IntegrationTests;
