+++ Snowflake.Data.Tests/IntegrationTests/SFDbCommandITAsync.cs
@@ -7,6 +7,7 @@ using Snowflake.Data.Core;
 using System.Linq;
 using System.IO;
 using Snowflake.Data.Telemetry;
+using Snowflake.Data.Tests.Util.Shims;
 
 namespace Snowflake.Data.Tests.IntegrationTests
 {
