+++ Snowflake.Data.Tests/Util/Shims/DbConnectionShims.cs
@@ -1,23 +1,14 @@
-using System.Collections.Generic;
+#if NETFRAMEWORK
 using System.Data;
 using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 using Snowflake.Data.Client;
 
-namespace Snowflake.Data.Tests.IntegrationTests;
+namespace Snowflake.Data.Tests.Util.Shims;
 
-#if NETFRAMEWORK || !NET8_0_OR_GREATER
-public static class FrameworkShims
+internal static class DbConnectionShims
 {
-    public static Task CancelAsync(this CancellationTokenSource cts)
-    {
-        cts.Cancel();
-        return Task.CompletedTask;
-    }
-
-#if NETFRAMEWORK
-
     public static Task<DbTransaction> BeginTransactionAsync(this SnowflakeDbConnection connection) => Task.FromResult(connection.BeginTransaction());
 
     public static Task CloseAsync(this DbConnection connection)
@@ -61,16 +52,5 @@ public static class FrameworkShims
         transaction.Rollback();
         return Task.CompletedTask;
     }
-
-    public static bool TryDequeue<T>(this Queue<T> queue, out T element)
-    {
-        element = default(T);
-        if (queue.Count == 0)
-            return false;
-
-        element = queue.Dequeue();
-        return true;
-    }
-#endif
 }
 #endif
